% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{citycostanalysis}
\alias{citycostanalysis}
\title{The Package}
\description{
This package provides a template to work with
shinythemes, dplyr, ggplot2, shiny, magriitr and integrated datasets.
}
