% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_city.R
\name{filter_city}
\alias{filter_city}
\title{City Filter}
\usage{
filter_city(Data)
}
\arguments{
\item{Data}{This function returns back the selected city for data}
}
\value{
Selected city to show cost of living
}
\description{
City Filter
}
