% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_total.R
\name{filter_total}
\alias{filter_total}
\title{Total Cost of living Calculator}
\usage{
filter_total(data)
}
\arguments{
\item{data}{Reflects back sum of Total Cost of living of the selected city in data table}
}
\value{
Sum of cost of living
}
\description{
Total Cost of living Calculator
}
