% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_data.R
\name{reactive_data}
\alias{reactive_data}
\title{Reactive Data}
\usage{
reactive_data(Data)
}
\arguments{
\item{this}{function returnd city input for toal cost of living for graph input}
}
\value{

}
\description{
Reactive Data
}
