% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{sampleGUI}
\alias{sampleGUI}
\title{Sample GUI}
\usage{
sampleGUI()
}
\value{
shiny app
}
\description{
Runs a Graphical User Interface to show comparison between cost of living among different major cities of the world.
}
